refactor: extract service collection extensions
