+++ backend/PhotoBank.Backend.sln
@@ -21,6 +21,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.ViewModel.Dto", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.BlobMigrator", "PhotoBank.BlobMigrator\PhotoBank.BlobMigrator.csproj", "{1265499D-2C5A-41FE-A82D-5E1F7EC22678}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.DependencyInjection", "PhotoBank.DependencyInjection\PhotoBank.DependencyInjection.csproj", "{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -139,6 +141,18 @@ Global
 		{1265499D-2C5A-41FE-A82D-5E1F7EC22678}.Release|x64.Build.0 = Release|Any CPU
 		{1265499D-2C5A-41FE-A82D-5E1F7EC22678}.Release|x86.ActiveCfg = Release|Any CPU
 		{1265499D-2C5A-41FE-A82D-5E1F7EC22678}.Release|x86.Build.0 = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|x64.Build.0 = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Debug|x86.Build.0 = Debug|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|x64.ActiveCfg = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|x64.Build.0 = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|x86.ActiveCfg = Release|Any CPU
+		{A80F3AE6-F138-4C41-BCFE-F1A24798FF67}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
