+++ backend/PhotoBank.Console/PhotoBank.Console.csproj
@@ -35,6 +35,7 @@
     <ProjectReference Include="..\PhotoBank.InsightFace.Client\PhotoBank.InsightFaceApiClient.csproj" />
     <ProjectReference Include="..\PhotoBank.Repositories\PhotoBank.Repositories.csproj" />
     <ProjectReference Include="..\PhotoBank.Services\PhotoBank.Services.csproj" />
+    <ProjectReference Include="..\PhotoBank.DependencyInjection\PhotoBank.DependencyInjection.csproj" />
   </ItemGroup>
 
   <ItemGroup>
