+++ backend/PhotoBank.sln
@@ -27,6 +27,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.InsightFaceApiCli
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.BlobMigrator", "PhotoBank.BlobMigrator\PhotoBank.BlobMigrator.csproj", "{794473DE-6A3C-45C2-A33A-D2E9767CFB61}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoBank.DependencyInjection", "PhotoBank.DependencyInjection\PhotoBank.DependencyInjection.csproj", "{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -185,6 +187,18 @@ Global
 		{794473DE-6A3C-45C2-A33A-D2E9767CFB61}.Release|x64.Build.0 = Release|Any CPU
 		{794473DE-6A3C-45C2-A33A-D2E9767CFB61}.Release|x86.ActiveCfg = Release|Any CPU
 		{794473DE-6A3C-45C2-A33A-D2E9767CFB61}.Release|x86.Build.0 = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|x64.Build.0 = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Debug|x86.Build.0 = Debug|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|x64.ActiveCfg = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|x64.Build.0 = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|x86.ActiveCfg = Release|Any CPU
+		{EB2353BC-7E4E-4F7B-B15D-DD38B50F6E22}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
